# server/swagger.yaml
openapi: 3.0.0
info:
  title: RapiDevKit
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /blocks/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Block
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Block
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
    post:
      summary: CRUD ACTION update
      tags:
        - Block
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Block"
    
  #-----  BlockType  -----
...
  /blocktypes/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - BlockType
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - BlockType
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockType"
    post:
      summary: CRUD ACTION update
      tags:
        - BlockType
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockType"
  /blocktypes/findByName/{key}:
    get:
      summary: CRUD ACTION findByName
      tags:
        - BlockType
      parameters:
        - name: key
          in: path
          description: Id of the resource Name to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockType"
    
  #-----  Item  -----
...
  /items/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    post:
      summary: CRUD ACTION update
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
  /items/findByType/{key}:
    get:
      summary: CRUD ACTION findByType
      tags:
        - Item
      parameters:
        - name: key
          in: path
          description: Id della risorsa Type da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/{id}/getLink:
    get:
      summary: CRUD ACTION getLink
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: ID of Item from Link
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION getLink
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: ID of Item from Link
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
    
  #-----  Role  -----
...
  /types/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Type
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Type
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Type"
    post:
      summary: CRUD ACTION update
      tags:
        - Type
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Type"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Type"
  /types/findByName/{key}:
    get:
      summary: CRUD ACTION findByName
      tags:
        - Type
      parameters:
        - name: key
          in: path
          description: Id of the resource Name to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Type"
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Block:
      properties:
        _id:
          type: string
        Actyve:
          type: boolean
        FontendAPIHook:
          type: string
      required:
        -  _id
        - Actyve
    BlockType:
      properties:
        _id:
          type: string
        Code:
          type: string
        Name:
          type: string
      required:
        -  _id
        - Name
    Item:
      properties:
        _id:
          type: string
        Active:
          type: boolean
        Created:
          type: integer
        Data:
          type: string
        Modified:
          type: integer
        Order:
          type: integer
        Protected:
          type: boolean
        Status:
          type: string
      required:
        -  _id
        - Active
...
    Type:
      properties:
        _id:
          type: string
        Name:
          type: string
        Schema:
          type: string
        Validation:
          type: string
        Visualization:
          type: string
      required:
        -  _id
...